Move AssetOrArchive to infer/types
