+++ infer/types/asset.go
